+++ src/vue/pages/dashboard/Dashboard.AssetSelector.vue
@@ -48,10 +48,10 @@
               <span class="asset-selector__asset-value-secondary">
                 {{
                   currentAssetBalanceDetails.convertedBalance | formatMoney({
-                    currency: config.DEFAULT_QUOTE_ASSET, symbolAllowed: true
+                    currency: defaultQuoteAsset, symbolAllowed: true
                   })
                 }}
-                {{ config.DEFAULT_QUOTE_ASSET }}
+                {{ defaultQuoteAsset }}
               </span>
             </div>
           </div>
@@ -100,8 +100,8 @@ export default {
   },
   props: {
     currentAsset: {
-      type: [String, Object],
-      default: config.DEFAULT_QUOTE_ASSET,
+      type: [String],
+      default: '',
     },
   },
   data: () => ({
@@ -114,6 +114,7 @@ export default {
   computed: {
     ...mapGetters({
       balances: vuexTypes.accountBalances,
+      defaultQuoteAsset: vuexTypes.defaultQuoteAsset,
     }),
     tokensList () {
       const balancesAssetCodes = this.balances.map(i => i.asset)